requirements.txt
build
Pillow
pytest
